pyproject.toml
[tool.poetry]
name = "m-django-extensions"
version = "0.1.2"
description = ""
authors = ["Artem <artem@dobroagency.ru>"]
readme = "README.md"
packages = [
    {include = "m_django_extensions"}
]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
